[DF-11108] Merge with refactored Reflection
Former-commit-id: 99f3c21a4d743bada78b7a191172cd9a23b5d992

+++ Sources/Internal/Engine/Private/Qt/PlatformCoreQt.cpp
@@ -33,8 +33,7 @@ void PlatformCore::Init()
 void PlatformCore::Run()
 {
     Vector<char*> qtCommandLine = engineBackend->GetCommandLineAsArgv();
-    int qtArgc = qtCommandLine.size();
-    ;
+    int qtArgc = static_cast<int>(qtCommandLine.size());
 
     QApplication app(qtArgc, qtCommandLine.data());
